+++ mullvad-types/src/settings/mod.rs
@@ -204,9 +204,10 @@ pub struct TunnelOptions {
     pub generic: GenericTunnelOptions,
     /// DNS options.
     pub dns_options: DnsOptions,
+    pub trusted_dns_options: TrustedDnsOptions,
 }
 
-pub use dns::{CustomDnsOptions, DefaultDnsOptions, DnsOptions, DnsState};
+pub use dns::{CustomDnsOptions, DefaultDnsOptions, DnsOptions, TrustedDnsOptions, DnsState};
 
 #[cfg(target_os = "android")]
 pub use dns::AndroidDnsOptions;
@@ -224,6 +225,7 @@ impl Default for TunnelOptions {
                 enable_ipv6: cfg!(target_os = "android"),
             },
             dns_options: DnsOptions::default(),
+            trusted_dns_options: TrustedDnsOptions::default(),
         }
     }
 }