+++ talpid-core/src/tunnel_state_machine/connecting_state.rs
@@ -354,6 +354,12 @@ impl ConnectingState {
                 Ok(_) => SameState(self.into()),
                 Err(cause) => self.disconnect(shared_values, AfterDisconnect::Block(cause)),
             },
+            Some(TunnelCommand::DnsTrust(servers)) => match shared_values.set_trusted_dns_servers(servers) {
+                #[cfg(target_os = "android")]
+                Ok(true) => self.disconnect(shared_values, AfterDisconnect::Reconnect(0)),
+                Ok(_) => SameState(self.into()),
+                Err(cause) => self.disconnect(shared_values, AfterDisconnect::Block(cause)),
+            },
             Some(TunnelCommand::BlockWhenDisconnected(block_when_disconnected)) => {
                 shared_values.block_when_disconnected = block_when_disconnected;
                 SameState(self.into())